change the casting on ej 3 for a f

+++ Tema1.1IntroduccionLenguajeJava/src/tipoDatosPrimitivos/Ej3.java
@@ -3,7 +3,7 @@ package tipoDatosPrimitivos;
 public class Ej3 {
 	
 	public void show () {
-		float f = (float) 9.87;
+		float f = 9.87f;
 		int i;
 		i = (int) f;
 		System.out.println(f);
